update with yellow zone

+++ cpp/core/utils/monitormmap/MetricReader.cpp
@@ -1,5 +1,12 @@
 /*
- * Copyright (c) Huawei Technologies Co., Ltd. 2025-2025. All rights reserved.
+ * Copyright (c) Huawei Technologies Co., Ltd. 2025. All rights reserved.
+ * You can use this software according to the terms and conditions of the Mulan PSL v2.
+ * You may obtain a copy of Mulan PSL v2 at:
+ *          http://license.coscl.org.cn/MulanPSL2
+ * THIS SOFTWARE IS PROVIDED ON AN "AS IS" BASIS, WITHOUT WARRANTIES OF ANY KIND,
+ * EITHER EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT,
+ * MERCHANTABILITY OR FIT FOR A PARTICULAR PURPOSE.
+ * See the Mulan PSL v2 for more details.
  */
 // MetricReader.cpp
 #include "MetricReader.h"
@@ -56,7 +63,7 @@ namespace omnistream {
             sharedMemoryFd = -1;
             return false;
         }
-        sharedMemorySize = shmStat.st_size;
+        sharedMemorySize = static_cast<size_t>(shmStat.st_size);
 
         sharedMemoryPtr = mmap(nullptr, sharedMemorySize, PROT_READ | PROT_WRITE, MAP_SHARED, sharedMemoryFd, 0);
         if (sharedMemoryPtr == MAP_FAILED) {
